[Refactor] setting of owner when requirement is created or edited/inspected; fixes GH142
Summary: when a requirement is created, set the owner to the domain-of-expertise that was selected when loggin in.

Test Plan: run unit test, run app and verify that when requiremet is created, the domain is set to the logged-in domain. When the requirement already exists, the owner should not be changed.

Reviewers: avorobiev

Reviewed By: avorobiev

Subscribers: naron.phou

Differential Revision: http://phabricator.rheagroup.com/D1502